move HealthCheck manifest structs to service file
